conversion and display of result

+++ src/main/java/de/sebastianschmelcher/currencyConverter/Model/ExchangeRates.java
@@ -1,4 +1,5 @@
 package de.sebastianschmelcher.currencyConverter.Model;
+import java.math.BigDecimal;
 import java.util.Map;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
@@ -7,7 +8,7 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 public class ExchangeRates {
 
 	private String base;
-	private Map<String,Double> rates;
+	private Map<String,BigDecimal> rates;
 	
     public String getBase() {
 		return base;
@@ -15,10 +16,10 @@ public class ExchangeRates {
 	public void setBase(String base) {
 		this.base = base;
 	}
-	public Map<String, Double> getRates() {
+	public Map<String, BigDecimal> getRates() {
 		return rates;
 	}
-	public void setRates(Map<String, Double> rates) {
+	public void setRates(Map<String, BigDecimal> rates) {
 		this.rates = rates;
 	}
 }
